// types/settings.ts
export interface ExternalModelItem {
  id: string;
  displayName: string;
  modelId: string;
}

export interface ExternalProvider {
  id: string;
  name: string;
  baseUrl: string;
  apiKey: string;
  models: ExternalModelItem[];
}